fix: [BCMS-926] timeout handling for wider compatibility (#127)
It re-applies #124 with proper conventions for commit messages (which allows automatic publish to run).

https://tiugotech.atlassian.net/browse/BCMS-926

* fix: BCMS-926 simplifies Butter instantiation

Removes the check for the `this` context during Butter class
instantiation, directly returning a new instance. This simplifies the
instantiation process, and should result in new instances each time this
is called in Node so that we respect user set timeouts

* fix: BCMS-926 timeout handling for React Native/Node.js

Addresses an issue where the timeout was not properly cleared in React
Native and Node.js environments when using a regular timeout (instead of
AbortSignal.timeout).

Updates the code to ensure the timeout is cleared by saving the timeout
ID and clearing it after use, preventing potential memory leaks or
unexpected behavior.

---------

Co-authored-by: Cobey Potter <cobey@wellfire.co>